requirements.txt
tensorflow
scipy
opencv-python
numpy
Pillow




#Train the model:
#python age_lsgan_transfer.py --gan_loss_weight=75 --fea_loss_weight=0.5e-4 --age_loss_weight=30 --fea_layer_name=conv5 --checkpoint_dir=./checkpoints/0_conv5_lsgan_transfer_g75_0.5f-4_a30/checkpoints/ --sample_dir=age/0_conv5_lsgan_transfer_g75_0.5f-4_a30 

#Train the model:
#python age_lsgan_transfer.py --gan_loss_weight=75 --fea_loss_weight=0.5e-4 --age_loss_weight=30 --fea_layer_name=fc6 --checkpoint_dir=./checkpoints/0_conv5_lsgan_transfer_g75_0.5f-4_a30/checkpoints_alejandro_v2/ --sample_dir=age/0_conv5_lsgan_transfer_g75_0.5f-4_a30 

#Test the model:
#python evaluation_metrics_v2.py --real_image_dir=./images/test/ --fake_image_dir=./age/0_conv5_lsgan_transfer_g75_0.5f-4_a30/test_ouput_finished